Deprecate usage of XML Spring schemas
Move SettingsNoMapping21 test

Related to #141.
